Allow skip of summarize handler during client setup. (#54)
<!-- Thank you for your contribution! Please provide the following
information -->
## Motivation and Context (Why the change? What's the scenario?)
We are seeing cases where the document summarization is being provided
instead of the relevant chunk:

```
0.79480314 {"file":"Building 34_35 Orientation.pdf","emb_src_file":"Building 34_35 Orientation.pdf.summarize.0.txt"
0.7932581 {"file":"Building 34_35 Orientation.pdf","emb_src_file":"Building 34_35 Orientation.pdf.partition.1.txt"
0.78576535 {"file":"Building 34_35 Orientation.pdf","emb_src_file":"Building 34_35 Orientation.pdf.partition.0.txt"
```

## High level description (Approach, Design)
Chat-Copilot would like the option to *not* introduce the document
summary until we can explore the feature further. At this point it only
servers to create an additional false-positive target with zero to none
added value.