Revisit IVectorDb interface (#175)
## Motivation and Context (Why the change? What's the scenario?)

The current `IVectorDb` interface assumes that embeddings are required
and must be generated when writing/reading from a DB. This is not always
true, e.g. when using Azure AI Search semantic ranking and Chroma DB
automatic embedding generation. The assumption is also limiting to a
single vector, while in theory one could use multiple and different
vector indexes on the same memory.

## High level description (Approach, Design)

* Rename `IVectorDb` to `IMemoryDb` to remove the assumption that one
must and can only use a Vector DB.
* Change `IVectorDb` read methods to receive the "text" (user input)
rather than the text embedding. Delegate to the implementing class the
decision whether to calculate embeddings and how. Currently use the
provided `ITextEmbeddingGeneration` to replicate the existing behavior
* Changes to configuration file used by the service
* Renamed some DI/builder methods to follow the new "memory db" naming
convention
* Fix some async code
* Added a dev only memory db not using embeddings, to show case the
scenario