Autodetect Docker CLI in Docker image build script (#299)
<!-- Thank you for your contribution! Please provide the following
information -->
## Motivation and Context (Why the change? What's the scenario?)

The Docker image build script only works with Docker CLI. Since in our
company we use ``podman`` instead of ``docker`` it would be nice if the
build script supports it.

## High level description (Approach, Design)

The script was modified to check which Docker compatible CLI is
installed in the machine. It checks if ``docker``, ``podman`` or
``nerdctl`` is available to build the image. Otherwise, it prints an
error.