Redis Missing Embedding score (#302)
## Motivation and Context (Why the change? What's the scenario?)

Fixes #298, Embedding score field was left out of the `RETURN` statement
added in #271 to prevent superfluous embeddings from coming back. When
`withEmbeddings` is set to true this is not a concern as it just pulls
the entire document in Redis back. When it's set to false we were
specifying which fields we wanted back, which did not include the
embedding score.

## High level description (Approach, Design)
Just need to append an extra field to the list of Return fields

---------

Co-authored-by: Devis Lucato <devis@microsoft.com>