Configure Qdrant default collection name (#339)
## Motivation and Context (Why the change? What's the scenario?)

As described in the issue
https://github.com/microsoft/kernel-memory/issues/285, it is useful to
have the ability to specify the default collection used by Qdrant, so we
don't need to specify it for every invocation of methos that work with
index.

## High level description (Approach, Design)

This PR adds a new configuration property, `DefaultIndex`, to
`QdrantConfig` class. This value is used to set the default index name
in the `QdrantMemory` client. If no value is provided, that the old
`Constants.DefaultIndex` field is used, so backward compatibility is
guaranted.

---------

Co-authored-by: Devis Lucato <devis@microsoft.com>