Reset stream position before posting to server (#509)
I am using the Kernel Memory project's MemoryWebClient to upload a stream to the server. The stream needs to be rewound before it is posted, but currently, this step seems to be missing, causing the server to receive an empty stream. When attempting to upload a stream using MemoryWebClient, the server receives an empty stream. It appears that the stream is not being rewound to its beginning before the upload, which results in the server not receiving the complete data as intended.

## High level description (Approach, Design)

Reset uploaded stream in MemoryWebclient

---------

Co-authored-by: David West <david.west@vistracorp.com>