Added initial implementation of Microsoft.KernelMemory.AI.ONNX package (#788)
## Motivation and Context (Why the change? What's the scenario?)

This change includes an entirely new extension that allows
TextGeneration with ONNX Models. This scenario allows users to interact
with local cloned ONNX Models

## High level description (Approach, Design)

This is a scenario that abstracts the general implementation of the
OnnxRuntimeGenAI library to work with text generation. More examples on
using OnnxGenAI Libraries, please see
https://github.com/microsoft/onnxruntime-genai

---------

Co-authored-by: Justin Ridings <juriding@microsoft.com>
Co-authored-by: Devis Lucato <devis@microsoft.com>